+++ game/g_local.h
@@ -1114,6 +1114,8 @@ extern	vmCvar_t	lmd_skillpoint_cost;
 extern	vmCvar_t	lmd_exp_for_level;
 extern	vmCvar_t    lmd_skillpoints_perlevel;
 extern	vmCvar_t    lmd_mastery_level_skills_level;
+extern	vmCvar_t	lmd_damage_level_scale;
+extern	vmCvar_t	lmd_is_lethality_add_damage;
 #include "../namespace_begin.h"
 
 // RoboPhred