+++ game/g_main.c
@@ -337,7 +337,9 @@ vmCvar_t	lmd_bounty_streaks_by;
 vmCvar_t	lmd_skillpoint_cost;
 vmCvar_t	lmd_exp_for_level;
 vmCvar_t    lmd_skillpoints_perlevel;
-vmCvar_t lmd_mastery_level_skills_level;
+vmCvar_t	lmd_mastery_level_skills_level;
+vmCvar_t	lmd_damage_level_scale;
+vmCvar_t	lmd_is_lethality_add_damage;
 								 
 //RoboPhred:
 vmCvar_t lmd_DataPath;
@@ -867,7 +869,8 @@ static cvarTable_t		gameCvarTable[] = {
 { &lmd_exp_for_level, "lmd_exp_for_level", "450", CVAR_ARCHIVE, 0, qfalse },
 { &lmd_skillpoints_perlevel, "lmd_skillpoints_perlevel", "4", CVAR_ARCHIVE, 0, qfalse },
 { &lmd_mastery_level_skills_level, "lmd_mastery_level_skills_level", "3", CVAR_ARCHIVE, 0, qfalse },
-
+{ &lmd_damage_level_scale, "lmd_damage_level_scale", "1", CVAR_ARCHIVE, 0, qfalse },
+{ &lmd_is_lethality_add_damage, "lmd_is_lethality_add_damage", "0", CVAR_ARCHIVE, 0, qfalse },
 
 // Lugormod cvars:
 { &g_noVoteTime, "g_noVoteTime", "5", CVAR_ARCHIVE,0, qfalse, qfalse,